debug mode
in the reforged version its not possible to use "debug mode" by enable it in the trigger editor menu. I added a global variable and renamed all "DEBUG_MODE" static variables to the new "IS_DEBUG_MODE" in the GameConfig.j. Just set this var to true/false to use debug mode